requirements.txt
streamlit==1.45.0
groq==0.25.0
plotly==6.0.1
pandas
numpy
python-dotenv==1.1.0
requests==2.32.3
folium==0.19.5
streamlit-folium==0.24.0
branca==0.8.1